feat: add package installation functionality
This commit has the beginnings of the package handling module. The
reason this is in the main workspace (image) is that alpm (arch linux
package management) supports operating within a chroot. This means that
there is no need to emulate the systemd-nspawn shell commands. This
should speed up image builds since we shouldn't need to emulate
architectures.