feat: initialize pacman keyring
First instance of shelling out to the pacman-key command since I
couldn't find a crate to handle that.

Also add comment in main.rs with a link to a forum post that says my
permissions might be messed up because I'm running this utility via sudo
and not as root. This seems to be corroborated from the install
instructions [1].

> Download and extract the root filesystem (as root, not via sudo):

[1]: https://archlinuxarm.org/platforms/armv8/broadcom/raspberry-pi-4